+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+July 17th, 2023
+Version 0.1.0
+- Improved challenge structure
+- Challenge 2 - Title Requestor
+- CTFd version 0.0.2 released
+
 July 15th, 2023
 Version 0.0.2
 - multithreading added to challenge 1 to support consecutive connections